+++ app/src/main/java/ru/gb/veber/paadlesson1/model/AppState.kt
@@ -1,5 +1,6 @@
-package ru.gb.veber.paadlesson1.model.data
+package ru.gb.veber.paadlesson1.model
 
+import ru.gb.veber.paadlesson1.model.datasources.network.DataModel
 
 sealed class AppState {
     data class Success(val data: List<DataModel>?) : AppState()
