--- app/src/main/java/ru/gb/veber/paadlesson1/model/data/Meanings.kt
@@ -1,8 +1,0 @@
-package ru.gb.veber.paadlesson1.model.data
-
-import com.google.gson.annotations.SerializedName
-
-class Meanings(
-    @field:SerializedName("translation") val translation: Translation?,
-    @field:SerializedName("imageUrl") val imageUrl: String?,
-)
