+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/DataSource.kt
@@ -1,7 +1,7 @@
 package ru.gb.veber.paadlesson1.model.datasources
 
 import io.reactivex.Observable
-
+// Источник данных для репозитория (Интернет, БД и т. п.)
 interface DataSource<T> {
     fun getData(word: String): Observable<T>
 }
