+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/DataSourceLocal.kt
@@ -1,7 +1,7 @@
 package ru.gb.veber.paadlesson1.model.datasources
 
 import io.reactivex.Observable
-import ru.gb.veber.paadlesson1.model.data.DataModel
+import ru.gb.veber.paadlesson1.model.datasources.network.DataModel
 
 class DataSourceLocal(
     private val remoteProvider: RoomDataBaseImplementation =
