+++ app/src/main/java/ru/gb/veber/paadlesson1/model/datasources/network/ApiService.kt
@@ -1,9 +1,8 @@
-package ru.gb.veber.paadlesson1.model.datasources
+package ru.gb.veber.paadlesson1.model.datasources.network
 
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Query
-import ru.gb.veber.paadlesson1.model.data.DataModel
 
 interface ApiService {
     @GET("words/search")
