+++ app/src/main/java/ru/gb/veber/paadlesson1/view/base/View.kt
@@ -1,6 +1,6 @@
 package ru.gb.veber.paadlesson1.view.base
 
-import ru.gb.veber.paadlesson1.model.data.AppState
+import ru.gb.veber.paadlesson1.model.AppState
 
 interface View {
     fun renderData(appState: AppState)
