+++ app/src/main/res/drawable/search_view_background.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/color_search_view_background"/>
+    <corners android:radius="15dp"/>
+</shape>
