+++ app/src/main/res/values/colors.xml
@@ -18,5 +18,6 @@
     <color name="color_black">#313131</color>
     <color name="color_white">#FFFFFF</color>
     <color name="color_gray">#66000000</color>
+    <color name="color_search_view_background">#eeeeee</color>
 
 </resources>
